requirements.txt
asn1crypto==1.0.1
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
coveralls==1.8.2
cryptography==2.7
docopt==0.6.2
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
